+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 1. Dashboard: Added ability to filter on applications and environments in Search Methods.
 2. Dashboard: Added ability to delete terminated agents.
 3. Agent: codekvast.conf: Added support for appVersion strategy "properties /path/to/file prop1,prop2"
+4. Drop database column users.lastActivityAt
 
 ## 0.23.1
 1. Database schema changes: applications, environments, jvms, invocations.